+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
 
+## 0.0.2
+- Added: SPFx debug extension configuration snippets
+- Added: Automatically adds the SPFx debug extension configuration when debug config gets initially created.
+
 ## 0.0.1
 - Initial release